CAY-2407 add prefetch to sqlTemplate in Modeler
