+++ RELEASE-NOTES.txt
@@ -23,6 +23,7 @@ CAY-2395 cdbimport: add option to create project file
 CAY-2396 Upgrade maven-assembly-plugin to 3.1.0
 CAY-2403 Extract eventbridges to top level
 CAY-2404 Move itests to maven-plugins
+CAY-2407 Modeler: add prefetch support for the SQLTemplate query
 
 Bug Fixes:
