+++ cayenne-server/src/main/java/org/apache/cayenne/query/SQLTemplate.java
@@ -491,6 +491,7 @@ public class SQLTemplate extends AbstractQuery implements ParameterizedQuery {
 		return metaData.getPrefetchTree();
 	}
 
+
 	/**
 	 * Adds a prefetch.
 	 *
