--- modeler/cayenne-modeler/src/main/java/org/apache/cayenne/modeler/editor/SQLTemplateOrderingTab.java
@@ -1,237 +1,0 @@
-/*****************************************************************
- *   Licensed to the Apache Software Foundation (ASF) under one
- *  or more contributor license agreements.  See the NOTICE file
- *  distributed with this work for additional information
- *  regarding copyright ownership.  The ASF licenses this file
- *  to you under the Apache License, Version 2.0 (the
- *  "License"); you may not use this file except in compliance
- *  with the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
- ****************************************************************/
-
-package org.apache.cayenne.modeler.editor;
-
-import org.apache.cayenne.configuration.event.QueryEvent;
-import org.apache.cayenne.map.Entity;
-import org.apache.cayenne.map.QueryDescriptor;
-import org.apache.cayenne.map.SQLTemplateDescriptor;
-import org.apache.cayenne.map.SelectQueryDescriptor;
-import org.apache.cayenne.modeler.Application;
-import org.apache.cayenne.modeler.ProjectController;
-import org.apache.cayenne.modeler.util.EntityTreeModel;
-import org.apache.cayenne.modeler.util.MultiColumnBrowser;
-import org.apache.cayenne.modeler.util.UIUtil;
-import org.apache.cayenne.query.Ordering;
-import org.apache.cayenne.query.SQLTemplate;
-import org.apache.cayenne.query.SortOrder;
-import org.apache.cayenne.swing.components.image.FilteredIconFactory;
-import org.apache.cayenne.util.CayenneMapEntry;
-
-import javax.swing.JLabel;
-import javax.swing.JPanel;
-import javax.swing.JScrollPane;
-import javax.swing.JSplitPane;
-import javax.swing.JTable;
-import javax.swing.ListSelectionModel;
-import javax.swing.event.ListSelectionEvent;
-import javax.swing.event.ListSelectionListener;
-import javax.swing.table.AbstractTableModel;
-import javax.swing.table.TableModel;
-import javax.swing.tree.TreeModel;
-import java.awt.BorderLayout;
-import java.awt.CardLayout;
-import java.awt.Dimension;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-import java.beans.PropertyChangeEvent;
-import java.beans.PropertyChangeListener;
-import java.util.prefs.Preferences;
-
-/**
- * A panel for picking SQLTemplate orderings.
- *
- */
-public class SQLTemplateOrderingTab extends JPanel implements PropertyChangeListener {
-    // property for split pane divider size
-    private static final String SPLIT_DIVIDER_LOCATION_PROPERTY = "query.orderings.divider.location";
-
-    static final Dimension BROWSER_CELL_DIM = new Dimension(150, 100);
-    static final Dimension TABLE_DIM = new Dimension(460, 60);
-
-    static final String PATH_HEADER = "Path";
-    static final String ASCENDING_HEADER = "Ascending";
-    static final String IGNORE_CASE_HEADER = "Ignore Case";
-
-    static final String REAL_PANEL = "real";
-    static final String PLACEHOLDER_PANEL = "placeholder";
-
-    protected ProjectController mediator;
-    protected SQLTemplateDescriptor sqlTemplate;
-
-    protected MultiColumnBrowser browser;
-    protected JTable table;
-
-    protected CardLayout cardLayout;
-    protected JPanel messagePanel;
-
-    public SQLTemplateOrderingTab(ProjectController mediator) {
-        this.mediator = mediator;
-
-        initView();
-        initController();
-    }
-
-    protected void initView() {
-
-        messagePanel = new JPanel(new BorderLayout());
-        cardLayout = new CardLayout();
-
-        Preferences detail = Application.getInstance().getPreferencesNode(this.getClass(), "");
-
-        int defLocation = Application.getFrame().getHeight() / 2;
-        int location = detail != null ? detail.getInt(
-                getDividerLocationProperty(),
-                defLocation) : defLocation;
-
-        //As of CAY-888 #3 main pane is now a JSplitPane. Top component is a bit larger.
-        JSplitPane mainPanel = new JSplitPane(JSplitPane.VERTICAL_SPLIT);
-        mainPanel.addPropertyChangeListener(JSplitPane.DIVIDER_LOCATION_PROPERTY, this);
-        mainPanel.setDividerLocation(location);
-
-        mainPanel.setTopComponent(createEditorPanel());
-        mainPanel.setBottomComponent(createSelectorPanel());
-
-        setLayout(cardLayout);
-        add(mainPanel, REAL_PANEL);
-        add(messagePanel, PLACEHOLDER_PANEL);
-    }
-
-    protected void initController() {
-
-        // scroll to selected row whenever a selection even occurs
-        table.getSelectionModel().addListSelectionListener(new ListSelectionListener() {
-
-            public void valueChanged(ListSelectionEvent e) {
-                if (!e.getValueIsAdjusting()) {
-                    UIUtil.scrollToSelectedRow(table);
-                }
-            }
-        });
-    }
-
-    protected void initFromModel() {
-        QueryDescriptor query = mediator.getCurrentQuery();
-
-        if (query == null || !QueryDescriptor.SQL_TEMPLATE.equals(query.getType())) {
-            processInvalidModel("Unknown query.");
-            return;
-        }
-
-        if (!(query.getRoot() instanceof Entity)) {
-            processInvalidModel("SQLTemplate has no root set.");
-            return;
-        }
-
-        this.sqlTemplate = (SQLTemplateDescriptor) query;
-        browser.setModel(createBrowserModel((Entity) sqlTemplate.getRoot()));
-
-        // init column sizes
-       // table.getColumnModel().getColumn(0).setPreferredWidth(250);
-
-        cardLayout.show(this, REAL_PANEL);
-    }
-
-    protected void processInvalidModel(String message) {
-        JLabel messageLabel = new JLabel(message, JLabel.CENTER);
-
-        messagePanel.removeAll();
-        messagePanel.add(messageLabel, BorderLayout.CENTER);
-        cardLayout.show(this, PLACEHOLDER_PANEL);
-    }
-
-    protected JPanel createEditorPanel() {
-
-        table = new JTable();
-        table.setRowHeight(25);
-        table.setRowMargin(3);
-        table.setPreferredScrollableViewportSize(TABLE_DIM);
-        table.setSelectionMode(ListSelectionModel.SINGLE_SELECTION);
-
-        JPanel panel = new JPanel(new BorderLayout());
-        panel.add(new JScrollPane(table), BorderLayout.CENTER);
-
-        return panel;
-    }
-
-    protected JPanel createSelectorPanel() {
-        browser = new MultiColumnBrowser();
-        browser.setPreferredColumnSize(BROWSER_CELL_DIM);
-        browser.setDefaultRenderer();
-
-        JPanel panel = new JPanel(new BorderLayout());
-        panel.add(new JScrollPane(
-                browser,
-                JScrollPane.VERTICAL_SCROLLBAR_NEVER,
-                JScrollPane.HORIZONTAL_SCROLLBAR_AS_NEEDED), BorderLayout.CENTER);
-
-        // setting minimal size, otherwise scrolling looks awful, because of
-        // VERTICAL_SCROLLBAR_NEVER strategy
-        panel.setMinimumSize(panel.getPreferredSize());
-
-        return panel;
-    }
-
-    protected TreeModel createBrowserModel(Entity entity) {
-        return new EntityTreeModel(entity);
-    }
-
-    protected String getSelectedPath() {
-        Object[] path = browser.getSelectionPath().getPath();
-
-        // first item in the path is Entity, so we must have
-        // at least two elements to constitute a valid ordering path
-        if (path.length < 2) {
-            return null;
-        }
-
-        StringBuilder buffer = new StringBuilder();
-
-        // attribute or relationships
-        CayenneMapEntry first = (CayenneMapEntry) path[1];
-        buffer.append(first.getName());
-
-        for (int i = 2; i < path.length; i++) {
-            CayenneMapEntry pathEntry = (CayenneMapEntry) path[i];
-            buffer.append(".").append(pathEntry.getName());
-        }
-
-        return buffer.toString();
-    }
-
-    /**
-     * Updates split pane divider location in properties
-     */
-    public void propertyChange(PropertyChangeEvent evt) {
-        if (JSplitPane.DIVIDER_LOCATION_PROPERTY.equals(evt.getPropertyName())) {
-            int value = (Integer) evt.getNewValue();
-
-            Preferences detail = Application.getInstance().getPreferencesNode(this.getClass(), "");
-            detail.putInt(getDividerLocationProperty(), value);
-        }
-    }
-
-    /**
-     * Returns name of a property for divider location.
-     */
-    protected String getDividerLocationProperty() {
-        return SPLIT_DIVIDER_LOCATION_PROPERTY;
-    }
-}
